Move update_check files to shared dir
